@@ -1,8 +1,10 @@
-const { Octokit } = require("@octokit/action");
-const core = require("@actions/core");
-const github = require('@actions/github');
-const { execSync } = require('child_process');
-const { existsSync } = require('fs');
+import { Octokit } from '@octokit/action';
+import * as core from '@actions/core';
+import * as github from '@actions/github';
+import { spawnSync } from 'child_process';
+import { existsSync } from 'fs';
+import { HttpsProxyAgent } from 'https-proxy-agent'
+import { getProxyForUrl } from 'proxy-from-env';
 
 const { runId, repo: { repo, owner }, eventName } = github.context;
 process.env.GITHUB_TOKEN = process.argv[2];
@@ -12,6 +14,12 @@ const lastSuccessfulEvent = process.argv[5];
 const workingDirectory = process.argv[6];
 const workflowId = process.argv[7];
 const defaultWorkingDirectory = '.';
+
+
+
+const ProxifiedClient = Octokit.plugin(
+  proxyPlugin
+);
 
 let BASE_SHA;
 (async () => {
@@ -24,10 +32,22 @@ let BASE_SHA;
     }
   }
 
-  const HEAD_SHA = execSync(`git rev-parse HEAD`, { encoding: 'utf-8' });
+  const headResult = spawnSync('git', ['rev-parse', 'HEAD'], { encoding: 'utf-8' });
+  const HEAD_SHA = headResult.stdout;
 
-  if (eventName === 'pull_request') {
-    BASE_SHA = execSync(`git merge-base origin/${mainBranchName} HEAD`, { encoding: 'utf-8' });
+
+  if (
+    (['pull_request', 'pull_request_target'].includes(eventName) && !github.context.payload.pull_request.merged) ||
+    eventName == 'merge_group'
+  ) {
+    try {
+      const mergeBaseRef = await findMergeBaseRef();
+      const baseResult = spawnSync('git', ['merge-base', `origin/${mainBranchName}`, mergeBaseRef], { encoding: 'utf-8' });
+      BASE_SHA = baseResult.stdout;
+    } catch (e) {
+      core.setFailed(e.message);
+      return;
+    }
   } else {
     try {
       BASE_SHA = await findSuccessfulCommit(workflowId, runId, owner, repo, mainBranchName, lastSuccessfulEvent);
@@ -47,7 +67,12 @@ let BASE_SHA;
         process.stdout.write('\n');
         process.stdout.write(`NOTE: You can instead make this a hard error by setting 'error-on-no-successful-workflow' on the action in your workflow.\n`);
 
-        BASE_SHA = execSync(`git rev-parse origin/${mainBranchName}~1`, { encoding: 'utf-8' });
+        const commitCountOutput = spawnSync('git', ['rev-list', '--count', `origin/${mainBranchName}`], { encoding: 'utf-8' }).stdout;
+        const commitCount = parseInt(stripNewLineEndings(commitCountOutput), 10);
+
+        const LAST_COMMIT_CMD = `origin/${mainBranchName}${commitCount > 1 ? '~1' : ''}`
+        const baseRes = spawnSync('git', ['rev-parse', LAST_COMMIT_CMD], { encoding: 'utf-8' });
+        BASE_SHA = baseRes.stdout;
         core.setOutput('noPreviousBuild', 'true');
       }
     } else {
@@ -55,9 +80,8 @@ let BASE_SHA;
       process.stdout.write(`Found the last successful workflow run on 'origin/${mainBranchName}'\n`);
       process.stdout.write(`Commit: ${BASE_SHA}\n`);
     }
-  }
 
-  const stripNewLineEndings = sha => sha.replace('\n', '');
+  }
   core.setOutput('base', stripNewLineEndings(BASE_SHA));
   core.setOutput('head', stripNewLineEndings(HEAD_SHA));
 })();
@@ -72,6 +96,15 @@ function reportFailure(branchName) {
     - If no, then you might have changed your git history and those commits no longer exist.`);
 }
 
+function proxyPlugin(octokit: Octokit) {
+  octokit.hook.before('request', options => {
+    const proxy: URL = getProxyForUrl(options.baseUrl)
+    if (proxy) {
+      options.request.agent = new HttpsProxyAgent(proxy)
+    }
+  })
+}
+
 /**
  * Find last successful workflow run on the repo
  * @param {string?} workflow_id
@@ -82,7 +115,7 @@ function reportFailure(branchName) {
  * @returns
  */
 async function findSuccessfulCommit(workflow_id, run_id, owner, repo, branch, lastSuccessfulEvent) {
-  const octokit = new Octokit();
+  const octokit = new ProxifiedClient();
   if (!workflow_id) {
     workflow_id = await octokit.request(`GET /repos/${owner}/${repo}/actions/runs/${run_id}`, {
       owner,
@@ -107,6 +140,33 @@ async function findSuccessfulCommit(workflow_id, run_id, owner, repo, branch, la
   return await findExistingCommit(shas);
 }
 
+async function findMergeBaseRef() {
+  if (eventName == 'merge_group') {
+    const mergeQueueBranch = await findMergeQueueBranch();
+    return `origin/${mergeQueueBranch}`;
+  } else {
+    return 'HEAD'
+  }
+}
+
+function findMergeQueuePr() {
+  const { head_ref, base_sha } = github.context.payload.merge_group;
+  const result = new RegExp(`^refs/heads/gh-readonly-queue/${mainBranchName}/pr-(\\d+)-${base_sha}$`).exec(head_ref);
+  return result ? result.at(1) : undefined;
+}
+
+async function findMergeQueueBranch() {
+  const pull_number = findMergeQueuePr();
+  if (!pull_number) {
+    throw new Error('Failed to determine PR number')
+  }
+  process.stdout.write('\n');
+  process.stdout.write(`Found PR #${pull_number} from merge queue branch\n`);
+  const octokit = new ProxifiedClient();
+  const result = await octokit.request(`GET /repos/${owner}/${repo}/pulls/${pull_number}`, { owner, repo, pull_number: +pull_number });
+  return result.data.head.ref;
+}
+
 /**
  * Get first existing commit
  * @param {string[]} commit_shas
@@ -128,9 +188,18 @@ async function findExistingCommit(shas) {
  */
 async function commitExists(commitSha) {
   try {
-    execSync(`git cat-file -e ${commitSha}`, { stdio: ['pipe', 'pipe', null] });
+    spawnSync('git', ['cat-file', '-e', commitSha], { stdio: ['pipe', 'pipe', null] });
     return true;
   } catch {
     return false;
   }
 }
+
+/**
+ * Strips LF line endings from given string
+ * @param {string} string
+ */
+function stripNewLineEndings(string) {
+  return string.replace('\n', '');
+}
+
